+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to BabelFish will be documented in this file.
 
 ## [1.12.2] - 2026-02-00
 ### Enhancements
+#### MatchSearchPublicRequest
+- Added ability to search for matches based on the owner of the match (aka Orion Club).
+#### ResultListIntermediateFormattedRow
+- Updated the return value for an Attribute to be the Field's Name (previously was the Field's Value).
 
 ### Bug Fixes
 #### MatchAbbr